+++ NEWS.md
@@ -4,6 +4,7 @@
 
 - RT cluster analysis for tentative isotope, adducts, and neutral loss peaks detection
 - Mass defect(cluster) analysis for homologous series detection
+- Add csv file generation for simulation data
 
 # enviGCMS 0.4.2