+++ NEWS.md
@@ -41,4 +41,5 @@
 - Add support for xcms 3 function
 
 ## Tips
+
 - Former related functions could be found in xsetplus package here: https://github.com/yufree/xsetplus